requirements.txt
numpy>=1.24.0
sounddevice>=0.4.6
openai-whisper>=20231117
